One Billion Corrhists (#197)
* add a billion corrhists

Bench: 12737644

* cargo fmt

* take 2, massively simpler

Bench: 12622269

+++ src/historytable.rs
@@ -230,8 +230,8 @@ impl CorrectionHistoryTable {
     }
 
     #[allow(clippy::cast_possible_truncation)]
-    pub fn get(&self, side: Colour, key: u64) -> i32 {
-        self.table[side][(key % CORRECTION_HISTORY_SIZE as u64) as usize]
+    pub fn get(&self, side: Colour, key: u64) -> i64 {
+        i64::from(self.table[side][(key % CORRECTION_HISTORY_SIZE as u64) as usize])
     }
 
     #[allow(clippy::cast_possible_truncation)]
